Sem7_CV_Project/CV_Project/parking_counter/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
ultralytics>=8.0.0
PyQt5>=5.15.0
PyYAML>=6.0
scipy>=1.11.0
Pillow>=10.0.0
torch==2.2.2+cpu
torchvision==0.17.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu